Fix: standardize competitionId types from Int to Long
Bug Risks (HIGH): Results module used Int for competitionId while other modules used Long, risking silent overflow for IDs exceeding Int.MAX_VALUE

- Added tests in CompetitionIdTypeConsistencyTest
- Changed competitionId to Long in ResultsRepository, ResultsDao, ResultsRemoteDataSource, Screen routes, AppNavHost navigation args, ResultsViewModel, ShooterResultViewModel, and all callers

+++ app/src/main/java/se/kjellstrand/webshooter/ui/screens/results/ResultsScreen.kt
@@ -278,7 +278,7 @@ fun CompetitionResultsScreen(
 @Composable
 fun ResultsList(
     resultsUiState: ResultsUiState,
-    competitionId: Int,
+    competitionId: Long,
     navController: NavController,
     resultsType: ResultsType = ResultsType.FIELD,
     listState: LazyListState = rememberLazyListState(),
@@ -350,7 +350,7 @@ fun ResultsList(
                                 navController.safeNavigate(
                                     Screen.ShooterResult.createRoute(
                                         competitionId,
-                                        result.signup.user.userID.toInt(),
+                                        result.signup.user.userID,
                                         resultsType.name
                                     )
                                 )
@@ -464,7 +464,7 @@ fun ResultsList(
                                     navController.safeNavigate(
                                         Screen.ShooterResult.createRoute(
                                             competitionId,
-                                            result.signup.user.userID.toInt(),
+                                            result.signup.user.userID,
                                             resultsType.name
                                         )
                                     )
